pose-extract gradientAscent(): distance interpolation, dynamic walkdir + step limit
2x2 interpolation allows us to estimate the distance more smoothly,
avoiding aliasing effects while determining the control points.

We now do not keep walkdir constant, but update it in each step as
if we started in that step. This leads to better guidance to the
center of the worm, but may also lead to oscillations, so we introduce
a limit on the maximum number of steps of the algorithm.