pose-extract: Extend the backbone to the tips of the worm (or image edge)
I was wondering whether this is worth the trouble as we will need to
account for translation (not just scale) when finding the zooming
transformation in the software using the pose anyway. However, we will
not have any very easy way to find how much away from the end of the
backbone to still look for interesting data, so by extending to tip we
provide a fixed frame of reference in this regard and we know we don't
have to look elsewhere.